isalpha() and isdigit() functions in C with example
isalpha(c) is a function in C which can be used to check if passed character is an alphabet or not. It returns a non-zero value if it’s an alphabet else it returns 0. For example, it returns non-zero values for ‘a’ to ‘z’ and ‘A’ to ‘Z’ and zero for other characters.

Similarly, isdigit(c) is a function in C which can be used to check if passed character is a digit or not. It returns a non-zero value if it’s a digit else it returns 0. For example, it returns non-zero value for ‘0’ to ‘9’ and zero for others.

Example Problem : Given a string str, find the number of alphabetic letters and number of decimal digits in that string.

Examples:

Input: 12abc12
Output: Alphabetic_letters = 3, Decimal_digits = 4

Input: 123 GeeksForGeeks is Number 1
Output: Alphabetic_letters = 21, Decimal_digits = 4