Pair : Simple Containers – The C++ Standard Template Library (STL)
The pair container is a simple container defined in <utility> header consisting of two data elements or objects.
The first element is referenced as ‘first’ and the second element as ‘second’ and the order is fixed (first, second).
Pair is used to combine together two values which may be different in type. Pair provides a way to store two heterogeneous objects as a single unit.
Pair can be assigned, copied and compared. The array of objects allocated in a map or hash_map are of type ‘pair’ by default in which all the ‘first’ elements are unique keys associated with their ‘second’ value objects.
To access the elements, we use variable name followed by dot operator followed by the keyword first or second.
Syntax :
 pair (data_type1, data_type2) Pair_name;


Note: Don't copy-paste the whole program as it contains different snippet of codes. Please go through them properly.